requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.6.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
python-multipart>=0.0.6
jinja2>=3.1.0
pandas>=2.0.0
openpyxl>=3.0.0
